requirements.txt
# Development dependencies
pytest>=8.0.0
black>=24.0.0
ruff>=0.1.0

# Project dependencies
google-adk>=1.22.1,<2.0.0
litellm>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
